# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Sereno API
  version: "1.0.0"
  description: >
    API do piloto de neuromodulação não invasiva. Contrato é a fonte de verdade.
    Erros seguem problem+json (RFC 9457). O braço do estudo nunca é exposto:
    endpoints de participante recebem apenas um handle de protocolo neutro, e a
    exportação de pesquisa usa braço codificado (A/B).
servers:
  - url: https://api.sereno.example/v1
security:
  - bearerAuth: []
paths:
...
  /screening/criteria:
    get:
      summary: Critérios de elegibilidade em vigor (staff)
      operationId: listScreeningCriteria
      description: >
        Catálogo versionado dos critérios do protocolo aprovado: chaves, rótulos e quais são
        DERIVADOS (calculados pelo servidor a partir dos escores, nunca declarados pelo
        formulário). Somente staff de inscrição (`enroll:write`).
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ScreeningCriteria' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /screening:
    post:
      summary: Triagem/elegibilidade do participante (staff) — ordena o funil de inscrição
      operationId: recordScreening
      description: >
        Registra a triagem: o servidor calcula a elegibilidade por regra transparente e
        versionada (todas as inclusões verdadeiras E nenhuma exclusão presente) e grava
        critérios + decisão. As chaves são as do protocolo aprovado e o conjunto é FECHADO —
        critério faltando, desconhecido ou derivado declarado é 422. É pré-condição, junto ao
        consentimento, para a alocação. Somente staff de inscrição (`enroll:write`).
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ScreeningIn' } } }
      responses:
        '201': { description: Triagem registrada, content: { application/json: { schema: { $ref: '#/components/schemas/ScreeningResult' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Problem:
      description: Erro (application/problem+json, RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [title, status]
...
    ScreeningIn:
      type: object
      description: >
        Critérios estruturados do protocolo aprovado. Elegível ⇔ todas as inclusões verdadeiras
        E nenhuma exclusão presente. `inclusion`/`exclusion` devem trazer o conjunto COMPLETO das
        chaves declaráveis (ver GET /screening/criteria); chave faltando, desconhecida ou
        derivada (`sintomas_elegiveis`, `d_gad7_grave_ou_risco`) é 422 — as derivadas o servidor
        calcula dos escores.
      properties:
        participant_id: { type: string, format: uuid }
        inclusion: { type: object, additionalProperties: { type: boolean } }
        exclusion: { type: object, additionalProperties: { type: boolean } }
        symptoms: { type: [object, "null"], additionalProperties: true }
        psqi_global:
          type: [integer, "null"]
          minimum: 0
          maximum: 21
          description: "PSQI global da triagem; > 5 satisfaz (com o GAD-7) a faixa sintomática de inclusão"
        phq9_items:
          type: [array, "null"]
          items: { type: integer, minimum: 0, maximum: 3 }
          minItems: 9
          maxItems: 9
          description: >-
            PHQ-9 de SEGURANÇA na triagem (não é desfecho). Item 9 positivo torna o candidato
            INELEGÍVEL e abre ficha de encaminhamento, como manda o protocolo.
        gad7_total:
          type: [integer, "null"]
          minimum: 0
          maximum: 21
          description: "GAD-7 da triagem; >= 15 é critério de exclusão e aciona encaminhamento"
      required: [participant_id, inclusion, exclusion]
    ScreeningResult:
      type: object
      properties:
        status: { type: string, enum: [screened] }
        eligible: { type: boolean }
        risk_detected: { type: boolean, description: "gatilho de segurança acionado na triagem" }
        unmet_criteria:
          type: array
          items: { type: string }
          description: "Chaves que barraram a inclusão (inclusão falsa ou exclusão presente)"
        referral_id: { type: [string, "null"], format: uuid }
      required: [status, eligible, risk_detected, unmet_criteria]
...
    ScreeningCriteria:
      type: object
      description: "Catálogo versionado dos critérios de elegibilidade em vigor."
      properties:
        version: { type: string }
        rule: { type: string }
        inclusion: { type: array, items: { $ref: '#/components/schemas/CriterionItem' } }
        exclusion: { type: array, items: { $ref: '#/components/schemas/CriterionItem' } }
        derived_from: { type: object, additionalProperties: { type: string } }
        note: { type: string }
      required: [version, rule, inclusion, exclusion]
    CriterionItem:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        derived: { type: boolean, description: "calculado pelo servidor; não declarar" }
      required: [key, label, derived]
...